Added the script file 'check-package-manager.sh'
In this commit, we have added the script file named check-package-manager.sh to the repository. The script serves the functions to determine the information of the package-manager installed on the linux distro (computer system) that the user is using currently. The script file is currently not filled with the required code, it only contains the commented documentation. The code will be updated soon.