Added the script file named check-distro.sh
In this commit, we have added the script file named  check-distro.sh to this repository / project. The script serves the functions to find out the information about the linux distribution that is currently installed in the user's computer system.  Although the script file currently contains the commented documentation and no code, but soon enough the file will be updated with error free and working code. ;-)