Added the code for recognizing the package-manager installed to 'check-package-manager.sh' script file
In this commit, we added all the code required to determine that which package manager software is currently installed on the computer system (Linux distro). The script can currently only determine limited package managers and they are : APT, PACMAN, DNF, RPM, YUM, and ABS. These package managers are only chosen because they use themselves as default and makes it much more easy for the user to execute the command even without automating. Like for updating the linux system, we can use the command  'apt update' if we are using the APT package manager. Slowly as per time, the script will be modified enough such that it determines even more package managers.